Move hooks to the new map directory
