+++ app/scripts/components/common/map/style-generators/cmr-timeseries.tsx
@@ -1,12 +1,9 @@
-import { useEffect, useMemo, useState } from 'react';
-import qs from 'qs';
-import { RasterSource, RasterLayer } from 'mapbox-gl';
+import React from 'react';
 
-import { requestQuickCache } from '../utils';
-import useMapStyle from '../hooks/use-map-style';
-import useGeneratorParams from '../hooks/use-generator-params';
 import { BaseGeneratorParams } from '../types';
-import { ActionStatus, S_FAILED, S_LOADING, S_SUCCEEDED } from '$utils/status';
+import { ZarrPaintLayer } from './zarr-timeseries';
+import { useCMR } from './hooks';
+import { ActionStatus } from '$utils/status';
 
 interface AssetUrlReplacement {
   from: string;
@@ -25,162 +22,17 @@ export interface CMRTimeseriesProps extends BaseGeneratorParams {
   onStatusChange?: (result: { status: ActionStatus; id: string }) => void;
 }
 
-const replaceInAssetUrl = (url: string, replacement: AssetUrlReplacement) => {
-  const {from, to } = replacement;
-  const cmrAssetUrl = url.replace(from, to);
-  return cmrAssetUrl;
-};
-
-export function CMRTimeseries(props: CMRTimeseriesProps) {
+export function CMRTimeseries(props:CMRTimeseriesProps) {
   const {
     id,
     stacCol,
     stacApiEndpoint,
-    tileApiEndpoint,
-    assetUrlReplacements,
     date,
-    sourceParams,
-    zoomExtent,
+    assetUrlReplacements,
     onStatusChange,
-    hidden,
-    opacity
   } = props;
 
-  const { updateStyle } = useMapStyle();
-  const [assetUrl, setAssetUrl] = useState('');
-
-  const [minZoom] = zoomExtent ?? [0, 20];
-
-  const stacApiEndpointToUse = stacApiEndpoint ?? process.env.API_STAC_ENDPOINT;
-
-  const generatorId = `zarr-timeseries-${id}`;
-
-  //
-  // Get the asset url
-  //
-  useEffect(() => {
-    const controller = new AbortController();
-
-    async function load() {
-      try {
-        onStatusChange?.({ status: S_LOADING, id });
-        if (!assetUrlReplacements) throw (new Error('CMR  layer requires asset url remplacement attributes'));
-
-        // Zarr collections in _VEDA_ should have a single entrypoint (zarr or virtual zarr / reference)
-        // CMR endpoints will be using individual items' assets, so we query for the asset url
-        const stacApiEndpointToUse = `${stacApiEndpoint}/search?collections=${stacCol}&datetime=${date?.toISOString()}`;
-
-        const data = await requestQuickCache({
-          url: stacApiEndpointToUse,
-          method: 'GET',
-          controller
-        });
-
-        const assetUrl =  replaceInAssetUrl((data as any).features[0].assets.data.href, assetUrlReplacements);
-        setAssetUrl(assetUrl);
-        onStatusChange?.({ status: S_SUCCEEDED, id });
-      } catch (error) {
-        if (!controller.signal.aborted) {
-          setAssetUrl('');
-          onStatusChange?.({ status: S_FAILED, id });
-        }
-        return;
-      }
-    }
-
-    load();
-
-    return () => {
-      controller.abort();
-    };
-  }, [id, stacCol, stacApiEndpointToUse, date, onStatusChange]);
-
-  //
-  // Generate Mapbox GL layers and sources for raster timeseries
-  //
-  const haveSourceParamsChanged = useMemo(
-    () => JSON.stringify(sourceParams),
-    [sourceParams]
-  );
-
-  const generatorParams = useGeneratorParams(props);
-
-  useEffect(
-    () => {
-      if (!tileApiEndpoint) return;
-
-      const tileParams = qs.stringify({
-        url: assetUrl,
-        time_slice: date,
-        ...sourceParams
-      });
-
-      const zarrSource: RasterSource = {
-        type: 'raster',
-        url: `${tileApiEndpoint}?${tileParams}`
-      };
-
-      const rasterOpacity = typeof opacity === 'number' ? opacity / 100 : 1;
-
-      const zarrLayer: RasterLayer = {
-        id: id,
-        type: 'raster',
-        source: id,
-        paint: {
-          'raster-opacity': hidden ? 0 : rasterOpacity,
-          'raster-opacity-transition': {
-            duration: 320
-          }
-        },
-        minzoom: minZoom,
-        metadata: {
-          layerOrderPosition: 'raster'
-        }
-      };
-
-      const sources = {
-        [id]: zarrSource
-      };
-      const layers = [zarrLayer];
-
-      updateStyle({
-        generatorId,
-        sources,
-        layers,
-        params: generatorParams
-      });
-    },
-    // sourceParams not included, but using a stringified version of it to
-    // detect changes (haveSourceParamsChanged)
-    [
-      updateStyle,
-      id,
-      date,
-      assetUrl,
-      minZoom,
-      tileApiEndpoint,
-      haveSourceParamsChanged,
-      generatorParams
-      // generatorParams includes hidden and opacity
-      // hidden,
-      // opacity,
-      // generatorId, // - dependent on id
-      // sourceParams, // tracked by haveSourceParamsChanged
-    ]
-  );
-
-  //
-  // Cleanup layers on unmount.
-  //
-  useEffect(() => {
-    return () => {
-      updateStyle({
-        generatorId,
-        sources: {},
-        layers: []
-      });
-    };
-  }, [updateStyle, generatorId]);
-
-  return null;
+  const stacApiEndpointToUse = stacApiEndpoint?? process.env.API_STAC_ENDPOINT;
+  const assetUrl = useCMR({ id, stacCol, stacApiEndpointToUse, date, assetUrlReplacements, stacApiEndpoint, onStatusChange });
+  return <ZarrPaintLayer {...props} assetUrl={assetUrl} />;
 }
